@@ -1,70 +1,70 @@
-�� p a c k a g e   c o m . z o o . m a i n ;  
-  
- i m p o r t   j a v a . i o . B u f f e r e d I n p u t S t r e a m ;  
- i m p o r t   j a v a . i o . I O E x c e p t i o n ;  
- i m p o r t   j a v a . i o . P r i n t W r i t e r ;  
- i m p o r t   j a v a . n e t . S e r v e r S o c k e t ;  
- i m p o r t   j a v a . n e t . S o c k e t ;  
-  
- p u b l i c   c l a s s   S e r v e r T e s t   {  
- 	 p u b l i c   s t a t i c   v o i d   m a i n ( S t r i n g [ ]   a r g s )   {  
- 	 	 S y s t e m . o u t . p r i n t l n ( "gR�z�T/R�� " ) ;  
- 	 	 S e r v e r S o c k e t   s s   =   n u l l ;  
- 	 	 S o c k e t   s   =   n u l l ;  
- 	 	  
- 	 	 / / I n p u t S t r e a m R e a d e r   i s r   =   n u l l ;  
- 	 	 / / B u f f e r e d R e a d e r   b r   =   n u l l ;  
- 	 	  
- 	 	 B u f f e r e d I n p u t S t r e a m   b i s   =   n u l l ;  
- 	 	 P r i n t W r i t e r   p w   =   n u l l ;  
- 	 	  
- 	 	 t r y   {  
- 	 	 	 s s   =   n e w   S e r v e r S o c k e t ( 8 8 8 8 ) ;  
- 	 	 	 s   =   s s . a c c e p t ( ) ;  
- 	 	 	  
- 	 	 	 b i s   =   n e w   B u f f e r e d I n p u t S t r e a m ( s . g e t I n p u t S t r e a m ( ) ) ;  
- 	 	 	 / / i s r   =   n e w   I n p u t S t r e a m R e a d e r ( b i s ) ;  
- 	 	 	 / / b r   =   n e w   B u f f e r e d R e a d e r ( i s r ) ;  
- 	 	 	  
- 	 	 	 i n t   x   =   b i s . r e a d ( ) ;  
- 	 	 	 w h i l e ( x ! =   - 1 )   {  
- 	 	 	 	 S y s t e m . o u t . p r i n t ( ( c h a r ) x ) ;  
- 	 	 	 	 x   =   ( i n t ) ( b i s . r e a d ( ) ) ;  
- 	 	 	 }  
- 	 	 	  
- 	 	 	 S y s t e m . o u t . p r i n t l n ( "gR�z�e6R0epcn^vQ�YSѐV�S�� " ) ;  
- 	 	 	 p w   =   n e w   P r i n t W r i t e r ( s . g e t O u t p u t S t r e a m ( ) ) ;  
- 	 	 	 p w . p r i n t ( " H e l l o C l i e n t " ) ;  
- 	 	 	 p w . f l u s h ( ) ;  
- 	 	 }   c a t c h   ( I O E x c e p t i o n   e )   {  
- 	 	 	 e . p r i n t S t a c k T r a c e ( ) ;  
- 	 	 }   f i n a l l y {  
- 	 	 	  
- 	 	 	 i f ( p w   ! =   n u l l )   {  
- 	 	 	 	 p w . c l o s e ( ) ;  
- 	 	 	 }  
- 	 	 	 i f ( b i s   ! =   n u l l )   {  
- 	 	 	 	 t r y   {  
- 	 	 	 	 	 b i s . c l o s e ( ) ;  
- 	 	 	 	 }   c a t c h   ( I O E x c e p t i o n   e )   {  
- 	 	 	 	 	 e . p r i n t S t a c k T r a c e ( ) ;  
- 	 	 	 	 }  
- 	 	 	 }  
- 	 	 	 i f (   s   ! =   n u l l )   {  
- 	 	 	 	 t r y   {  
- 	 	 	 	 	 s . c l o s e ( ) ;  
- 	 	 	 	 }   c a t c h   ( I O E x c e p t i o n   e )   {  
- 	 	 	 	 	 e . p r i n t S t a c k T r a c e ( ) ;  
- 	 	 	 	 }  
- 	 	 	 }  
- 	 	 	 i f ( s s   ! =   n u l l )   {  
- 	 	 	 	 t r y   {  
- 	 	 	 	 	 s s . c l o s e ( ) ;  
- 	 	 	 	 }   c a t c h   ( I O E x c e p t i o n   e )   {  
- 	 	 	 	 	 e . p r i n t S t a c k T r a c e ( ) ;  
- 	 	 	 	 }  
- 	 	 	 }  
- 	 	 	  
- 	 	 }  
- 	 }  
- }  
+package com.zoo.main;
+
+import java.io.BufferedInputStream;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.net.ServerSocket;
+import java.net.Socket;
+
+public class ServerTest {
+	public static void main(String[] args) {
+		System.out.println("服务端启动！");
+		ServerSocket ss = null;
+		Socket s = null;
+
+		//InputStreamReader isr = null;
+		//BufferedReader br = null;
+
+		BufferedInputStream bis = null;
+		PrintWriter pw = null;
+
+		try {
+			ss = new ServerSocket(8888);
+			s = ss.accept();
+
+			bis = new BufferedInputStream(s.getInputStream());
+			//isr = new InputStreamReader(bis);
+			//br = new BufferedReader(isr);
+
+			int x = bis.read();
+			while(x!= -1) {
+				System.out.print((char)x);
+				x = (int)(bis.read());
+			}
+
+			System.out.println("服务端收到数据并准备发送回去！");
+			pw = new PrintWriter(s.getOutputStream());
+			pw.print("HelloClient");
+			pw.flush();
+		} catch (IOException e) {
+			e.printStackTrace();
+		} finally{
+
+			if(pw != null) {
+				pw.close();
+			}
+			if(bis != null) {
+				try {
+					bis.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
+			if( s != null) {
+				try {
+					s.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
+			if(ss != null) {
+				try {
+					ss.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
+
+		}
+	}
+}
